Update auth library
- Use ActivityResultLauncher rather than registering for ActivityResult. This moves the post-login logic from MainActivityViewModel to WelcomeScreenViewModel.
- Also moved tests
- Include logic in WelcomeScreen to handle onwards routing

Resolves: DCMAW-9390
