+++ app/build.gradle.kts
@@ -178,6 +178,7 @@ dependencies {
         libs.network,
         libs.jose4j,
         projects.features,
+        libs.runtime.livedata,
         platform(libs.firebase.bom),
         libs.firebase.appcheck,
         libs.firebase.appcheck.debug
