+++ app/src/main/java/uk/gov/onelogin/login/ui/welcome/WelcomeScreen.kt
@@ -1,8 +1,13 @@
 package uk.gov.onelogin.login.ui.welcome
 
+import android.app.Activity
+import androidx.activity.compose.rememberLauncherForActivityResult
+import androidx.activity.result.ActivityResult
+import androidx.activity.result.contract.ActivityResultContracts
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.ui.platform.LocalContext
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.livedata.observeAsState
 import androidx.hilt.navigation.compose.hiltViewModel
 import uk.gov.android.onelogin.R
 import uk.gov.android.ui.components.content.GdsContentText
@@ -15,9 +20,23 @@ fun WelcomeScreen(
     viewModel: WelcomeScreenViewModel = hiltViewModel(),
     navigateToOfflineErrorScreen: () -> Unit = { },
     shouldTryAgain: () -> Boolean = { false },
-    openDeveloperPanel: () -> Unit = { }
+    openDeveloperPanel: () -> Unit = { },
+    navigatePostLogin: (String) -> Unit = { }
 ) {
-    val context = LocalContext.current
+    val next by viewModel.next.observeAsState()
+    next?.let {
+        navigatePostLogin(it)
+    }
+
+    val launcher = rememberLauncherForActivityResult(
+        ActivityResultContracts.StartActivityForResult()
+    ) { result: ActivityResult ->
+        if (result.resultCode == Activity.RESULT_OK) {
+            result.data?.let { intent ->
+                viewModel.handleActivityResult(intent = intent)
+            }
+        }
+    }
     LandingPage(
         landingPageParameters =
         LandingPageParameters(
@@ -32,7 +51,7 @@ fun WelcomeScreen(
             ),
             onPrimary = {
                 if (viewModel.onlineChecker.isOnline()) {
-                    viewModel.onPrimary(context)
+                    viewModel.onPrimary(launcher)
                 } else {
                     navigateToOfflineErrorScreen()
                 }
@@ -52,7 +71,7 @@ fun WelcomeScreen(
         if (!shouldTryAgain()) return@LaunchedEffect
 
         if (viewModel.onlineChecker.isOnline()) {
-            viewModel.onPrimary(context)
+            viewModel.onPrimary(launcher)
         } else {
             navigateToOfflineErrorScreen()
         }
