+++ gradle/libs.versions.toml
@@ -45,8 +45,10 @@ googleServices = "4.4.2"
 firebaseBom = "33.1.2"
 classgraph = "4.8.174"
 govuk-ui = "4.4.1"
+compose-runtime = "1.6.8"
 
 [libraries]
+runtime-livedata = { group = "androidx.compose.runtime", name = "runtime-livedata", version.ref = "compose-runtime" }
 jose4j = { module = "org.bitbucket.b_c:jose4j", version.ref = "jose4j" }
 components = "uk.gov.android:components:4.4.1"
 test-core-ktx = { group = "androidx.test", name = "core-ktx", version = "1.5.0" }
@@ -69,7 +71,7 @@ pages = { group = "uk.gov.android", name = "pages", version.ref = "govuk-ui" }
 slf4j-api = { group = "org.slf4j", name = "slf4j-api", version = "2.0.12" }
 theme = { group = "uk.gov.android", name = "theme", version.ref = "govuk-ui" }
 uiautomator = { group = "androidx.test.uiautomator", name = "uiautomator", version = "2.2.0" }
-authentication = "uk.gov.android.authentication:app:0.6.0"
+authentication = "uk.gov.android.authentication:app:0.6.2"
 androidx-core-ktx = { group = "androidx.core", name = "core-ktx", version.ref = "core-ktx" }
 androidx-test-ext-junit = { group = "androidx.test.ext", name = "junit", version.ref = "androidx-test-ext-junit" }
 androidx-espresso-core = { group = "androidx.test.espresso", name = "espresso-core", version.ref = "espresso-core" }
